angular.app/stark/packages/stark-core/src/validation/decorators/is-kbo/is-kbo.validator.decorator.spec.ts
/*tslint:disable:completed-docs*/
import { validateSync, ValidationError } from "class-validator";
import { StarkIsKBO } from "./is-kbo.validator.decorator";
import { starkIsKBOValidatorName } from "../../validators/is-kbo";

class MyClass {
	@StarkIsKBO()
	public name: string;
}

class SimpleClass {
	@StarkIsKBO()
	public dummyObject: object;
}

describe("ValidatorDecorator: StarkIsKBO", () => {
	let myClass: MyClass;
	let simpleClass: SimpleClass;
	const validatorConstraintName: string = starkIsKBOValidatorName;

	beforeEach(() => {
		myClass = new MyClass();
		simpleClass = new SimpleClass();
	});

	it("should fail if the object to validate is not a string", () => {
		const errors: ValidationError[] = validateSync(simpleClass);

		expect(errors.length).toBe(1);
		expect(errors[0].constraints).toBeDefined();
		expect(errors[0].constraints[validatorConstraintName]).toBeDefined();
	});

	it("should fail if kbo is empty", () => {
		const errors: ValidationError[] = validateSync(myClass);

		expect(errors.length).toBe(1);
		expect(errors[0].constraints).toBeDefined();
		expect(errors[0].constraints[validatorConstraintName]).toBeDefined();
	});

	it("should fail if kbo is not correct", () => {
		myClass.name = "2.104.471.666";
		const errors: ValidationError[] = validateSync(myClass);

		expect(errors.length).toBe(1);
	});

	it("should NOT fail if kbo is correct", () => {
		myClass.name = "2.104.471.814";
		const errors: ValidationError[] = validateSync(myClass);

		expect(errors.length).toBe(0);
	});
});
